Now calling move for a ball computes the next position of the ball for the next frame.

+++ elements_painter.py
@@ -17,12 +17,12 @@ def paint_ball(ball, surface):
     surface : the destination surface => PyGame.Surface
     """
     pygame.draw.circle(surface, ball.get_bg_color(), ball.get_center(),
-                       ball.get_diameter() / 2)
+                       int(ball.get_diameter() / 2))
     ball_center = ball.get_center()
     txt_width, txt_height = ball.get_txt_font().size(ball.get_operation().
                                                      get_text())
-    txt_position = (ball_center[0] - txt_width / 2,
-                    ball_center[1] - txt_height / 2)
+    txt_position = (int(ball_center[0] - txt_width / 2),
+                    int(ball_center[1] - txt_height / 2))
     txt_surface = ball.get_txt_font().render(
         ball.get_operation().get_text(), color=ball.get_txt_color())
     surface.blit(txt_surface, txt_position)
